Test instability fixing - ServiceUnavailableException: HTTP 503 Service Unavailable
It happens there is thrown exception

javax.ws.rs.ServiceUnavailableException: HTTP 503 Service Unavailable
    at org.jboss.resteasy.client.jaxrs.internal.ClientInvocation.handleErrorStatus(ClientInvocation.java:191)
    at org.jboss.resteasy.client.jaxrs.internal.ClientInvocation.extractResult(ClientInvocation.java:154)
    at org.jboss.resteasy.client.jaxrs.internal.ClientInvocation.invoke(ClientInvocation.java:444)
    at org.jboss.narayana.rts.TxnHelper.sendRequest(TxnHelper.java:87)
    at org.jboss.narayana.rts.TxnTest.testTxn(TxnTest.java:66)

which seems to be an intermittent failure. It seems the timeout of 20s
could not be enough for processing enlistment of a resource at the
transaction manager during the performance load and the tests starts to
fail.