Bumping Undertow version used in jta tests to 2.0.16
It was a try to fix an intermittent failures with
error

javax.ws.rs.ServiceUnavailableException: HTTP 503 Service Unavailable
    at org.jboss.resteasy.client.jaxrs.internal.ClientInvocation.handleErrorStatus(ClientInvocation.java:191)
    at org.jboss.resteasy.client.jaxrs.internal.ClientInvocation.extractResult(ClientInvocation.java:154)
    at org.jboss.resteasy.client.jaxrs.internal.ClientInvocation.invoke(ClientInvocation.java:444)
    at org.jboss.narayana.rts.TxnHelper.sendRequest(TxnHelper.java:87)
    at org.jboss.narayana.rts.TxnTest.testTxn(TxnTest.java:66)

was caused by Undertow stopped to listen on the test port 8090. The
reason was not found but sounds the new version of Undertow does not
have this trouble.

The issue was moved to jira as https://issues.jboss.org/browse/JBTM-3092